feat!: Allow applications to pick their C_x
The change allows applications to pick usable C_x (i.e., C_I and C_R)
values, which they are in a position to decide, because unlike the EDHOC
library, they keep track of all the ongoing exchanges.

The c_wrapper API is *not* changed at this point, because the API change
would be way too subtle (the "out" parameter would be changed to an "in"
parameter).

BREAKING CHANGE: This alters the message 1 and 2 API.