fix!: Make edhoc.rs generic over crypto implementation
This is likely not a final version, as it is terribly verbose. The
verbosity would be mitigated if the functions were methods of their
states, as then, it would be concentrated atop the impl blocks.

It is also incomplete in that it merely shifts the cfg-switched Crypto
implementation from being referenced in edhdoc.rs to being referenced in
lib.rs, instead of making the latter generic as well.