refactor: Typestating of edhoc.rs, removal of error codes
Primarily, this changes the State structs of edhoc.rs to use type state.
Unlike for the lib.rs interface in previous commits, this uses generics
and a marker trait instead of standalone types, mainly because there is
only little to be gained by adding and removing fields over time as they
become populated or consumed. (That change can also be made later with
way less refactoring impact).

As a result, functions that could only return the WrongState error (eg.
the edhoc_exporter function) now are infallible, and the WrongState
error variant is removed.

Changes to some of the functions led to minor rewrites (where State was
stored in the same mutable variable for its in and out form previously,
which is not possible any more here) and the Success error state could
be removed from them. That state was ever only used in locally mutable
errors, and never returned. The Success state was removed from the
EDHOCError enum, and all other functions that relied on mutable local
errors or states