refactor!: Pull Crypto trait to the high-level implementation
This allows pushing back the edhoc-crypto ("the default implementation
that is selected statically, making all implementations possible
dependencies") into the dev-dependencies.

The crypto-* features are removed from edhoc-rs; testing depends on
edhoc-crypto being pulled in in parallel to the test, and a feature
selected on that.

Follow-up-for: https://github.com/openwsn-berkeley/edhoc-rs/pull/127