refactor: Apply suggestions from clippy
The changes were hand-screened for whether they make sense in this
particular case; most did. Unused variables were not renamed (because
they may indicate a lack of further processing that would warrant more
detailed investigation); value-in-range checks were not applied and are
considered more of a style decision.

Several assign-and-return pairs were applied. While the original forms
do contribute a bit to readability when used with proper names (eg.
th_2), most cases just used `output`, so all are changed for
consistency. (And the functions are usually short enough that the name
`compute_X` is directly visible anyway).