python: Don't return Vec<u8>
The ID_CRED_x returned in the initiator and responder APIs are
unintuitive for Python users when arrays (esp. if they see examples
where it's a `[10]` style single value -- to a Python user that looks
more like a parsed CBOR, more so if that is actually what parsing a CBOR
sequence would give). Returning it as Python bytes makes it easier to
understand that this is the opaque data indicating either an identifier
or a credential by value.

Closes: https://github.com/openwsn-berkeley/lakers/issues/256