crypto: Take slices and return generic arrays
This reduces the amount of necessary copying, and decouples the
functions from predefined lengths (given that they were originally
designed for a concrete type, then used with other types, which only
works while those are aliased).

The length generic being decoupled from the input length does raise a
few new cases of panicking (maybe even unsoundness, in the cryptocell
case), but at least for encrypting (where more data is written to the
output than in the input), those were already present before, and are
better tackled independently from the refactoring.

Contributes-To: https://github.com/openwsn-berkeley/lakers/issues/363