buffers: Align types against mismatching type aliases
The new types are all aliases of the old, but those now work even when
of all the type aliases of EdhocMessageBuffer in shared/src/lib.rs, new
explicitly different sizes are chosen. (Except for EdhocCiphertext2 and
EdhocPlaintext2 which need to be the same size lest their
encrypt_decrypt function needs duplication).