core: Avoid manual population of buffers, push/extend_from_slice instead
This simultaneously avoids going through direct accessors, and reduces
the risk of errors by using the `.len` field as a cursor. (The added
calculations do not necessarily cause overhead: the compiler can know
that it has exclusive access, run the sum in a register, and write it at
the end).